% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{affy}
\alias{affy}
\title{Greene-5P01NS017771-220003: Microarray expression experiment related with Parkinson disease.}
\source{
GEO Dataset from NCBI, GEO accession: GSE4773.
}
\description{
AffyBatch object, has a part of raw expression values obtained from microarray chip.
}
\author{
Greene JG, 2006.
}
