% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCP.R
\name{CCP}
\alias{CCP}
\title{Finding Common Connection Pattern between different networks.}
\usage{
CCP(...)
}
\arguments{
\item{...}{The networks (igraph objects) to obtain the Common Connection Patterns.}
}
\value{
An igraph object with all the Common Connection Pattern in the same network.
}
\description{
From the intersection of two or more networks, it obtains the connected components 
by deleting the solitary nodes in the intersection network. They must be graph objects.
}
\examples{
# Loading data

data("net1")
data("net2")

# Obtaining Common Connection Patterns

ccp <- CCP(net1,net2)
ccp
}
\author{
Juan David Henao <judhenaosa@unal.edu.co>
}
