% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNet.R
\name{createNet}
\alias{createNet}
\title{Creating a co-expression network from expression matrix.}
\usage{
createNet(expData, method, threshold)
}
\arguments{
\item{expData}{A whole expression matrix or differentially expressed genes matrix, it may be stored in a SummarizedExperiment object.}

\item{method}{A function to calculate the similarity matrix between genes. It can be "correlation" to use Pearson function or "mutual information" to use a based on entropy information function.}

\item{threshold}{A value between 0 and 1 to filter the similarity matrix and create the co-expression network.}
}
\value{
An undirected co-expression network as igraph object.
}
\description{
From an expression matrix, this function creates a co-expression network like a graph object using a threshold value and one similarity function.
}
\examples{

# Loading data

pathfile <- system.file("extdata","expression_example.txt",package = "coexnet")
data <- read.table(pathfile,stringsAsFactors = FALSE)

# Building the network

cor_pearson <- createNet(expData = data,threshold = 0.7,method = "correlation")
cor_pearson

mut_inf <- createNet(expData = data,threshold = 0.5,method = "mutual information")
mut_inf
}
\seealso{
\code{\link{findThreshold}} to obtain a threshold value based on biology network assumptions.
}
\author{
Juan David Henao Sanchez <judhenaosa@unal.edu.co>

Liliana Lopez Kleine <llopezk@unal.edu.co>
}
