% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif.exprs.R
\name{dif.exprs}
\alias{dif.exprs}
\title{Differential expression analysis using two different methods.}
\usage{
dif.exprs(eset, treatment, fdr, DifferentialMethod)
}
\arguments{
\item{eset}{A data.frame with the expression matrix.}

\item{treatment}{A vector with the ientifiers of the classes, 0 to control
and 1 to case.}

\item{fdr}{The expected FDR value.}

\item{DifferentialMethod}{The method to calculate the differential expressed
genes, can be "sam" or "acde"}
}
\value{
A data.frame with the expression matrix to the expressed diferential 
genes only.
}
\description{
Using the expression matrix calculate the differential expressed
genes to two class analysis and fixing an expected FDR value. The methods
are SAM and ACDE.
}
\author{
Juan David Henao <judhenaosa@unal.edu.co>
}
\references{
Tusher, V. G., Tibshirani, R., & Chu, G. (2001). Significance analysis of 
microarrays applied to the ionizing radiation response. Proceedings of the National Academy of Sciences, 98(9), 5116-5121.

Acosta J and Lopez-Kleine L (2015). acde: Artificial Components Detection of Differentially Expressed Genes. R package version 1.4.0.
}
\seealso{
\code{\link{expr.mat}} to obtain the expression matrix.
}

