% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr.mat.R
\name{expr.mat}
\alias{expr.mat}
\title{Calculate the expression matrix from the raw expression data.}
\usage{
expr.mat(affy, genes, NormalizeMethod, SummaryMethod, BatchCorrect = TRUE)
}
\arguments{
\item{affy}{A AffyBatch object with the raw expression data.}

\item{genes}{A table with two columns, in the firt one the name of each probe 
in the microarray with header "probe" and in the second one the corresponding gene or ID with header "ID".}

\item{NormalizeMethod}{The method to normalize the raw data. Can be "vsn" to apply 
Variance Stabilizing Normalization function or "rma" to apply Robust Multi-Array Average function.}

\item{SummaryMethod}{The method to pass from probeset to genes or ID. Can be "max" to selecto the probeset with the most
average expression value or "median" to obtain the median of each sample of the set of probeset corresponding to particular
gen or ID.}

\item{BatchCorrect}{The option to apply batch effect correction, by default TRUE.}
}
\value{
A expression matrix, in the rows each one of the gene/ID and in the columns each one of the samples.
}
\description{
This function use a affyBatch object with the raw expression data to normalize and transform the matrix from
probeset to gene considering the option to remove the batch effect in the long microarray data.
}
\examples{
\dontrun{

# Loading AffyBatch object

data("affy")

# Loading table with probeset and gene/ID information

data("info")

# Calculating the expression matrix

## RMA

rma <- expr.mat(affy = affy,genes = info,NormalizeMethod = "rma",
SummaryMethod = "median",BatchCorrect = FALSE)
head(rma)

## VSN

vsn <- expr.mat(affy = affy,genes = info,NormalizeMethod = "vsn",
SummaryMethod = "median",BatchCorrect = FALSE)
head(vsn)

}
}
\author{
Juan David Henao <judhenaosa@unal.edu.co>
}
\references{
Huber, W., Von Heydebreck, A., Sultmann, H., Poustka, A., & Vingron, M. (2002). Variance stabilization applied to microarray data calibration and to the quantification of differential expression. Bioinformatics, 18(suppl 1), S96-S104.

Irizarry, R. A., Hobbs, B., Collin, F., Beazer Barclay, Y. D., Antonellis, K. J., Scherf, U., & Speed, T. P. (2003). 
Exploration, normalization, and summaries of high density oligonucleotide array probe level data. Biostatistics, 4(2), 249-264.
}
\seealso{
\code{\link{get.affy}} to obtain the affyBatch object.

\code{\link{gene.symbol}} to obtain the data.frame with probeset and genes/ID from .SOFT file.
}

