% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.threshold.R
\name{find.threshold}
\alias{find.threshold}
\title{Find the threshold value to create a co-expression network}
\usage{
find.threshold(difexp, method)
}
\arguments{
\item{difexp}{The data.frame with the expression matrix for the expressed differential genes.}

\item{method}{The method to create the correlation matrix, can be "correlation" to Pearson test or "mutual information" to test 
based on entropy information.}
}
\value{
The best threshold value find using the to criteria and a plot with the result.
}
\description{
Find the threshold value to establish the edges in a co-expression network based on correlation matrix
using two different criteria. First, use the differential value between the clustering coefficient for the real network and
random network build and second test the network using a Kolgomorov-Smirnov test to reject normality in the degree distribution of 
the edges in the network.
}
\author{
Juan David Henao Sanchez <judhenaosa@unal.edu.co>
}
\seealso{
\code{\link{dif.exprs}} to find the differential expressed genes matrix.
}

