% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findThreshold.R
\name{findThreshold}
\alias{findThreshold}
\title{Find the threshold value to create a co-expression network}
\usage{
findThreshold(expData, method, plotting = FALSE)
}
\arguments{
\item{expData}{A whole expression matrix or the expression matrix to differentially expressed genes, it may be stored in a SummarizedExperiment object.}

\item{method}{The method name to create the correlation matrix, this can be "correlation" to obtain the Pearson Correlation Coefficient. On the other hand, can be "mutual information" to obtain the correlation values from an entropy-based method.}

\item{plotting}{The option to show the result in a plot. By default FALSE.}
}
\value{
The best threshold value found using the two criteria and a plot showing the result.
}
\description{
Finds the threshold value to establish the cutoff in the process to define the edges in the co-expression network final from two steps. In the first one, obtains the subtraction from clustering coefficient values of the real and random networks created from the possible threshold values in the correlation matrix. In the second one, a Kolmogorov-Smirnov test is made to evaluate the degree distribution respect normality.
}
\examples{

# Loading data

pathfile <- system.file("extdata","expression_example.txt",package = "coexnet")
data <- read.table(pathfile,stringsAsFactors = FALSE)

# Finding threshold value

cor_pearson <- findThreshold(expData = data,method = "correlation")
cor_pearson
}
\references{
Elo, L. L., Jarvenpaa, H., Oresic, M., Lahesmaa, R., & Aittokallio, T. (2007). Systematic construction of gene coexpression networks with applications to human T helper cell differentiation process. Bioinformatics, 23(16), 2096-2103.

Leal, L. G., Lopez, C., & Lopez-Kleine, L. (2014). Construction and comparison of gene co-expression networks shows complex plant immune responses. PeerJ, 2, e610.
}
\seealso{
\code{\link{difExprs}} to find the differentially expressed genes matrix.
}
\author{
Juan David Henao Sanchez <judhenaosa@unal.edu.co>

Liliana Lopez Kleine <llopezk@unal.edu.co>
}
