% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.symbol.R
\name{gene.symbol}
\alias{gene.symbol}
\title{Create a table relating probesets with genes.}
\usage{
gene.symbol(GPL, d = ".")
}
\arguments{
\item{GPL}{The GPL ID.}

\item{d}{The pathfile where the .soft file are, by default is the current pathfile.}
}
\value{
A data.frame with two columns, in the first one the probesets and in the second one the corresponding gene to each probeset.
}
\description{
From the information in the .soft file, creates a data.frame with two columns, the first one with the probeset names
and the second one with the name of the corresponding genes. 
The genes are annotated using the gene symbol system.
}
\author{
Juan David Henao <judhenaosa@unal.edu.co>
}

