% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.info.R
\name{get.info}
\alias{get.info}
\title{Download raw expression data from GEO DataSet}
\usage{
get.info(GSE, GPL, dir = ".")
}
\arguments{
\item{GSE}{The list of GSE ID.}

\item{GPL}{The GPL ID of the chip used in the microarray experiment.}

\item{dir}{The pathfile to download and uncompress the raw expression data, by default is the current pathfile.}
}
\value{
A set of files with the uncompressed data and ready to use.
}
\description{
Using a GSE ID list, from the GEO DtaSet ftp, search and download each one of the raw expression data in a 
separate file with the GSE ID as the name, after that, this function uncompress the downloaded file to be ready to use.
Adittionally, this function search and downloaded the information of the chip using the GPL ID to obtain the .soft file.
}
\examples{

# Extract data from GEO DataSet

get.info(GSE = "GSE8216", GPL = "GPL2025",dir = tempdir())
}
\author{
Juan David Henao <judhenaosa@unal.edu.co>
}
\seealso{
\code{\link{get.affy}} to charge the expression data.
}

