% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAffy.R
\name{getAffy}
\alias{getAffy}
\title{Charge and create an AffyBatch object}
\usage{
getAffy(GSE, directory = ".")
}
\arguments{
\item{GSE}{The name of the file with the compressed samples data.}

\item{directory}{The path file where the samples and the "filelist" file are to create the AffyBatch object, by default is the current directory.}
}
\value{
An AffyBatch Object.
}
\description{
Charges the data from a file with the GSM samples compressed using the "filelist.txt" file to identify the GSM ID
and create the AffyBatch object finally.
}
\examples{

# Creating the AffyBatch object from raw data downloaded
# The data is from partial experiment results from Greene JG(2006), GEO accession: GSE4773

# Data without CDF environment information

affy <- getAffy(GSE = "GSE1234",directory = system.file("extdata",package = "coexnet"))
}
\seealso{
\code{\link{getInfo}} to download expression data from GEO DataSets included the "filelist" file.
}
\author{
Juan david Henao <judhenaosa@unal.edu.co>
}
