% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInfo.R
\name{getInfo}
\alias{getInfo}
\title{Download raw expression data from GEO DataSet}
\usage{
getInfo(GSE, GPL, directory = ".")
}
\arguments{
\item{GSE}{The list of GSE ID.}

\item{GPL}{The GPL ID of the chip used in the microarray experiment.}

\item{directory}{The path file to download and uncompress the raw expression data, by default is the current path file.}
}
\value{
A set of files with the uncompressed data and ready to use.
}
\description{
Using a GSE ID list, from the GEO DataSets FTP, it searches and downloads each raw expression data in a separate file with the GSE ID as the name. After, the function uncompresses the downloaded file to be ready to use. Additionally, searches and downloaded the information of the chip using the GPL ID to obtain the .soft file.
}
\examples{
\dontrun{
# Extract data from GEO DataSets (Takes time)

getInfo(GSE = "GSE8216", GPL = "GPL2025",directory = tempdir())
}
}
\seealso{
\code{\link{getAffy}} to charge the expression data.
}
\author{
Juan David Henao <judhenaosa@unal.edu.co>
}
