% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{info}
\alias{info}
\title{Table with the relation between the probesets and the corresponding gene.}
\source{
GEO Dataset from NCBI, GPL accession: GPL8300.
}
\description{
Table where the probesets and genes are related.
}
\author{
Greene JG, 2006.
}
