% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppiNet.R
\name{ppiNet}
\alias{ppiNet}
\title{Create a protein-protein interaction network}
\usage{
ppiNet(molecularIDs = NULL, file = NULL, speciesID = 9606,
  evidence = c("neighborhood", "neighborhood_transferred", "fusion",
  "cooccurence", "homology", "coexpression", "coexpression_transferred",
  "experiments", "experiments_transferred", "database", "database_transferred",
  "textmining", "textmining_transferred", "combined_score"))
}
\arguments{
\item{molecularIDs}{A vector of IDs recognized by STRING database to create a PPI network from them.}

\item{file}{A file with an edge list to charge the PPI network.}

\item{speciesID}{The numerical ID from STRING database to the species of interest, by defect, is "9006" corresponding to human species.}

\item{evidence}{A vector with the evidence to support the interactions between the proteins, by default is all the evidence given in STRING database.}
}
\value{
An igraph object as the protein-protein interaction network where the nodes are the molecular identifiers given in the input.
}
\description{
Creates a protein-protein interaction network using an edge list with the relations between proteins or a vector with the gene symbols or any other molecular identifier type, widely used in biological databases, to create a predictive PPI network using information of evidence in STRING database.
}
\examples{
\dontrun{
# Creating a vector with identifiers

ID <- c("FN1","HAMP","ILK","MIF","NME1","PROCR","RAC1","RBBP7",
"TMEM176A","TUBG1","UBC","VKORC1")

# Creating the PPI network

ppi <- ppiNet(molecularIDs = ID,evidence = c("neighborhood","coexpression","experiments"))
ppi
}

# Creating a PPI network from external data

ppi <- ppiNet(file = system.file("extdata","ppi.txt",package = "coexnet"))
ppi
}
\author{
Juan David Henao <judhenaosa@unal.edu.co>
}
