% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.components.R
\name{shared.components}
\alias{shared.components}
\title{Finding shared components between diferent networks.}
\usage{
shared.components(...)
}
\arguments{
\item{...}{The networks (igraph objects) to obtain the shared components.}
}
\value{
A vector with the names of the shared components.
}
\description{
From the intersected network, obtain the nodes without any value of degree. There are
the molecular shared components without biological relation among them.
}
\author{
Juan David Henao <judhenaosa@unal.edu.co>
}
