% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCP.R
\name{CCP}
\alias{CCP}
\title{Finding Common Connection Pattern between different networks.}
\usage{
CCP(...)
}
\arguments{
\item{...}{The networks (igraph objects) to obtain the Common Connection Patterns.}
}
\value{
An igraph object with all the Common Connection Pattern in the same network.
}
\description{
From the ntersection of two or more networks, obntain the connected components deleting
the alone nodes in the intersection network. The networks must be igraph objects.
}
\examples{
# Charge data

data("net1")
data("net2")

# Obtain Common Connection Patterns

ccp <- CCP(net1,net2)
ccp
}
\author{
Juan David Henao <judhenaosa@unal.edu.co>
}
