% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.net.R
\name{create.net}
\alias{create.net}
\title{Create a co-expression matrix using correlation matrix}
\usage{
create.net(difexp, method, threshold)
}
\arguments{
\item{difexp}{A data.frame with the expressed diferential genes matrix.}

\item{method}{The method to create the correlation matrix, can be "correlation" to use the Pearson Correlation Coefficient or
"mutual information" to use the Mutual Information test.}

\item{threshold}{The threshold value to establish the edges in the co-expression network.}
}
\value{
A igraph objecto with the co-expression network.
}
\description{
Create a co-expression network from the expressed diferential genes matrix and a threshold value using a 
correlation matrix created from two different methods. The first one is the Pearson Correlation Coefficient and the second one is 
the Mutual Information test.
}
\author{
Juan David Henao Sanchez <judhenaosa@unal.edu.co>
}
\seealso{
\code{\link{find.threshold}} to find a threshold value.
}

