% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.net.R
\name{create.net}
\alias{create.net}
\title{Create a co-expression network from expression matrix.}
\usage{
create.net(difexp, method, threshold)
}
\arguments{
\item{difexp}{A whole expression matrix or the expression matrix to differentially expressed genes.}

\item{method}{The function to calculate the similarity among the genes. "correlation" to use Pearson function or 
"mutual information" to use a function based on entropy information.}

\item{threshold}{A value (between 0.1 to 0.99) to cut off the adjacency matrix and create the co-expression network.}
}
\value{
A undirected co-expression network like igraph object.
}
\description{
From a expression matrix, this function create a co-expression network like a igrph object using a threshold value and
one similarity function.
}
\examples{

# Loading data

pathfile <- system.file("extdata","expression_example.txt",package = "coexnet")
data <- read.table(pathfile,stringsAsFactors = FALSE)

# Building the network

cor_pearson <- create.net(difexp = a,threshold = 0.7,method = "correlation")
cor_pearson

mut_inf <- create.net(difexp = a,threshold = 0.5,method = "mutual information")
mut_inf
}
\author{
Juan David Henao Sanchez <judhenaosa@unal.edu.co>

Liliana Lopez Kleine <llopezk@unal.edu.co>
}
\seealso{
\code{\link{find.threshold}} to obtain a threshold value based on biology network assumptions.
}

