% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.affy.R
\name{get.affy}
\alias{get.affy}
\title{Charge and create an AffyBatch object}
\usage{
get.affy(GSE, dir = ".")
}
\arguments{
\item{GSE}{The name of the file with the compressed samples data.}

\item{dir}{The path file where the samples and the "filelist" file are to create the AffyBatch object, by default is the current directory.}
}
\description{
Charges the data from a file with the GSM samples compressed using the "filelist.txt" file to identify the GSM ID
and create the AffyBatch object finally.
}
\examples{

# Creating the AffyBatch object from raw data downloaded
# The data is from partial experiment results from Greene JG(2006), GEO accession: GSE4773

# Data without CDF environment information

affy <- get.affy(GSE = "GSE1234",dir = system.file("extdata",package = "coexnet"))
}
\seealso{
\code{\link{get.info}} to download expression data from GEO DataSets included the "filelist" file.
}
\author{
Juan david Henao <judhenaosa@unal.edu.co>

# Bioinformatics and Systems Biology | Universidad Nacional de Colombia
}
