% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.affy.R
\name{get.affy}
\alias{get.affy}
\title{Charge and create an AffyBatch object}
\usage{
get.affy(GSE)
}
\arguments{
\item{GSE}{The name of the file with the compressed samples data.}
}
\description{
Charges the data from a file with the GSM samples compressed using the "filelist.txt" file to identify the GSM ID
and create the AffyBatch object finally.
}
\examples{

# Load the AffyBatch from downloaded raw data

affy <- get.affy("GSE8216")
}
\author{
Juan david Henao <judhenaosa@unal.edu.co>
}
\seealso{
\code{\link{get.info}} to download expression data from GEO DataSet.
}

