% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.components.R
\name{shared.components}
\alias{shared.components}
\title{Finding shared components between diferent networks.}
\usage{
shared.components(...)
}
\arguments{
\item{...}{The networks (igraph objects) to obtain the shared components.}
}
\value{
A vector with the names of the shared components.
}
\description{
From the intersection network, obtains the nodes without any degree value. These are the shared molecular components without any biological relation among them.
}
\examples{
# Loading data

data("net1")
data("net2")

# Obtaining shared components

share <- shared.components(net1,net2)
share
}
\author{
Juan David Henao <judhenaosa@unal.edu.co>
}
